+++ srcs/vm/statements/and.c
@@ -3,23 +3,24 @@
 /*                                                        :::      ::::::::   */
 /*   and.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ajones <ajones@student.42.fr>              +#+  +:+       +#+        */
+/*   By: egaliber <egaliber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/02 12:47:25 by sam               #+#    #+#             */
-/*   Updated: 2023/03/17 17:01:51 by ajones           ###   ########.fr       */
+/*   Updated: 2023/03/19 16:04:20 by egaliber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vm.h"
 
 /*
-* -
+* - Performs a "bitwise AND" statement for the values​of the first two arguments.
+* - and writes the result to the registry passed as the third argument.
 */
 void	ft_and(t_carriage *carriage)
 {
-	uint8_t value_1;
-	uint8_t value_2;
-	uint8_t result;
+	int	value_1;
+	int	value_2;
+	int	result;
 
 	value_1 = fetch_value(carriage, &carriage->args[0]);
 	value_2 = fetch_value(carriage, &carriage->args[1]);
