+++ srcs/vm/statements/xor.c
@@ -3,23 +3,23 @@
 /*                                                        :::      ::::::::   */
 /*   xor.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ajones <ajones@student.42.fr>              +#+  +:+       +#+        */
+/*   By: egaliber <egaliber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/02 12:47:53 by sam               #+#    #+#             */
-/*   Updated: 2023/03/17 17:04:24 by ajones           ###   ########.fr       */
+/*   Updated: 2023/03/19 16:21:00 by egaliber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vm.h"
-
 /*
-* -
+* - Xor is almost identical to the <and> statement.
+* - Only difference is, "bitwise AND" is replaced by "bitwise exclusive OR".
 */
 void	ft_xor(t_carriage *carriage)
 {
-	uint8_t value_1;
-	uint8_t value_2;
-	uint8_t result;
+	int	value_1;
+	int	value_2;
+	int	result;
 
 	value_1 = fetch_value(carriage, &carriage->args[0]);
 	value_2 = fetch_value(carriage, &carriage->args[1]);
